Add draft Assignment 2 (EC2 section only)
This draft only covers the EC2 section of the assignment.  The Katacoda
Git scenarios must still be added.

This draft also assumes that `aws-cred.sh` is used for setting the AWS
credentials for the CLI and the CMPT 756 tools.  If the instructor
wishes to use a different method, such as calls to `aws configure` and
direct editing of `cluster/tpl-vars.txt`, this draft will have to be updated.