Use `app-a*.py` source file names
The recent commit to `c756-exer' changed the source file names
from every exercise using 'app.py' to using 'app-aN.py', where N is
the assignment number.

This commit updates A1 and A2 to use the new convention. The page
for A3 already used the new convention and so did not require updating.